Rename deviations section and reorganize related files
The "deviation section" was a bit of a misnomer
because it only deals with the "missing routes" deviations,
when actually there are other deviations as well.
So, move the general deviations title stuff to a more suitable place.
