+++ ui/src/hooks/timetables-import/deviations/index.ts
@@ -1,9 +1,6 @@
 export * from './duplicate-journeys';
+export * from './missing-routes';
 export * from './useCreateVehicleScheduleFrameInfo';
-export * from './useDeviationSort';
-export * from './useFindOrphanRoutes';
 export * from './useGetStagingVehicleScheduleFrameIds';
-export * from './useReplaceDeviations';
-export * from './useToReplaceVehicleScheduleFrames';
 export * from './useVehicleScheduleFrameWithJourneys';
 export * from './useVehicleScheduleFrameWithRouteLabelAndLineId';
