+++ ui/src/hooks/timetables-import/deviations/missing-routes/useFindOrphanRoutes.spec.ts
@@ -1,6 +1,6 @@
 import { buildLocalizedString } from '@hsl/jore4-test-db-manager';
-import { RouteDirectionEnum } from '../../../generated/graphql';
-import { renderHook } from '../../../utils/test-utils/renderer';
+import { RouteDirectionEnum } from '../../../../generated/graphql';
+import { renderHook } from '../../../../utils/test-utils/renderer';
 import { useFindOrphanRoutes } from './useFindOrphanRoutes';
 
 const notUsedPartOfInfoObject = {
